面试邀请接口相关

+++ seaman-recruit/src/main/java/com/lhh/seamanrecruit/entity/UserPosition.java
@@ -49,7 +49,7 @@ public class UserPosition implements Serializable {
 	 * 投递标识
 	 */
 	@ApiModelProperty(value = "投递标识(0-未投递,1-已投递,2-通过,3-拒绝)")
-	private Long deliveryFlag;
+	private Integer deliveryFlag;
 
 	/**
 	 * 创建时间
